Restaurant and Menus Added
Added list of restaurants and menu of Banana Joes

+++ UCSC_Meal_Builder/app/src/main/java/com/example/sam/ucsc_meal_builder/Item.java
@@ -6,9 +6,9 @@ package com.example.sam.ucsc_meal_builder;
 public class Item {
 
     String name;
-    float price;
+    double price;
 
-    public Item(String n, float p){
+    public Item(String n, double p){
         name = n;
         price = p;
 
@@ -19,7 +19,7 @@ public class Item {
     }
 
     public String toString(){
-        return name + " " + Float.toString(price);
+        return name + " " + Double.toString(price);
     }
 
 }
